test(scripts): extract markdown table-cell escaper and cover it
scripts/summarize-job-source-check.mjs escaped its job-source table cells with a
local escapeTableCell (backslash/pipe escaping plus whitespace-run collapsing)
that was only covered indirectly through formatJobSourceCheckSummary. Move it into
scripts/lib/table-cell.mjs - matching the existing scripts/lib convention - and
import it back, adding direct tests.

- scripts/lib/table-cell.mjs: escapeTableCell.
- scripts/summarize-job-source-check.mjs: import the escaper; drop the local copy.
- tests: cover pipe/backslash escaping (backslash first), collapsing each run of
  spaces/tabs/newlines to a single space, trimming, and nullish/non-string
  coercion. Branch coverage 100% (10/10).

No behavior change; the rendered table cells are identical.
